Move solver from TaskDataSheet one level up
I added a new abstract concept, datasheet, which the task-data-sheet
and the future sot-data-sheet should exptend. In the process, the name
is loaded to this new abstract concept.

Moreover, the solvers concept is also pushed one level up to the
abstract concept so that other sot datahseet will aslo share it.