Add package.xml and provide option to create stream in ops/cpp
Package.xml is added mainly because how crappy is the catkin
mechanisms. This done just to fix the issue of the header fiels which
in fact is only optional. That is, we can now create streams for
rtt_ros in either the cpp or in ops. The reason that i kept cpp as an
option is to ensure that in case ops comes from other places, we still
have all set.